add controller

+++ Models/Data/Money.cs
@@ -3,9 +3,9 @@ using System.Runtime.Serialization;
 namespace PersonalWebApi.Models.Data;
 
 [DataContract]
-public class Money(int amount, string? currencyCode) {
+public class Money(decimal amount, string? currencyCode) {
     [DataMember]
-    public int Amount { get; private set; } = amount;
+    public decimal Amount { get; private set; } = amount;
     [DataMember]
     public string CurrencyCode { get; private set; } = currencyCode ?? "JPY";
 }
